Update SurfaceCpdf.php
The php-svg-lib is not compatible with PHP 7.4 for SVGs using transparency. Within the library, the currentLineTransparency and currentFillTransparency properties from the CPdf class are set to null. This class is dependent on the existence of both properties as array. In PHP 7.4 this will throw an error.

Initial testing does not show any issues without unsetting the properties, although this still needs proper testing. It might affect performance as the return statement in Cpdf.php:3790 is not hit anymore, where this might have been the case in previous PHP versions.